refactot: change dinamic view current distribution
