--- client/src/store/packageIdSelected.tsx
@@ -1,15 +1,0 @@
-import { createSlice } from '@reduxjs/toolkit';
-
-const packageSlice = createSlice({
-  name: 'package',
-  initialState: '',
-  reducers: {
-    setPackage: (state, action) => {
-      return action.payload;
-    },
-  },
-});
-
-export const { setPackage } = packageSlice.actions;
-
-export default packageSlice.reducer;
