+++ client/src/styles/CurrentDistribution.module.css
@@ -18,6 +18,13 @@
   flex-direction: column;
 }
 
+.loading {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 18px;
+}
+
 .accordion_space {
   margin-bottom: 10%;
 }
